Distinguish mediated/non-mediated connection state in NodeJS, rename files
