+++ agents/node/vcxagent-core/package.json
@@ -35,7 +35,7 @@
     "demo:faber:mysql": "node demo/faber.js --mysql",
     "demo:alice:sign": "node demo/alice-signature.js",
     "demo:faber:verify": "node demo/faber-verify-signature.js",
-    "test:integration": "npm run test:integration:issue-verify && npm run test:integration:signing && npm run test:integration:messaging && npm run test:integration:tails && npm run test:integration:trustping && npm run test:integration:feature-discovery && npm run test:integration:public-invite && npm run test:integration:out-of-band && npm run test:integration:nonmediated-endpoint && npm run test:integration:nonmediated-connection",
+    "test:integration": "npm run test:integration:issue-verify && npm run test:integration:signing && npm run test:integration:messaging && npm run test:integration:tails && npm run test:integration:trustping && npm run test:integration:feature-discovery && npm run test:integration:public-invite && npm run test:integration:out-of-band && npm run test:integration:endpoint && npm run test:integration:connection",
     "test:integration:issue-verify": "jest --forceExit --env=node --runInBand test/issue-verify.spec.js",
     "test:integration:signing": "jest --forceExit --env=node --runInBand test/sign-verify.spec.js",
     "test:integration:messaging": "jest --forceExit --env=node --runInBand test/messaging.spec.js",
@@ -44,8 +44,8 @@
     "test:integration:feature-discovery": "jest --forceExit --env=node --runInBand test/feature-discovery.spec.js",
     "test:integration:public-invite": "jest --forceExit --env=node --runInBand test/public-invite.spec.js",
     "test:integration:out-of-band": "jest --forceExit --env=node --runInBand test/out-of-band.spec.js",
-    "test:integration:nonmediated-endpoint": "jest --forceExit --env=node --runInBand test/nonmediated-endpoint.spec.js",
-    "test:integration:nonmediated-connection": "jest --forceExit --env=node --runInBand test/nonmediated-connection.spec.js"
+    "test:integration:endpoint": "jest --forceExit --env=node --runInBand test/endpoint.spec.js",
+    "test:integration:connection": "jest --forceExit --env=node --runInBand test/connection.spec.js"
   },
   "dependencies": {
     "@hyperledger/node-vcx-wrapper": "file:../../../wrappers/node",
