+++ agents/node/vcxagent-core/src/agent.js
@@ -9,7 +9,7 @@ const { createServiceProver } = require('./services/service-prover')
 const { createServiceCredHolder } = require('./services/service-cred-holder')
 const { createServiceCredIssuer } = require('./services/service-cred-issuer')
 const { createServiceConnections } = require('./services/service-connections')
-const { createServiceNonmediatedConnections } = require('./services/service-nonmediated-connections')
+const { createServiceNonmediatedConnections } = require('./services/service-mediated-connections')
 const { createServiceOutOfBand } = require('./services/service-out-of-band')
 const { createServiceLedgerRevocationRegistry } = require('./services/service-revocation-registry')
 const { provisionAgentInAgency } = require('./utils/vcx-workflows')
